requirements.txt
# Runtime dependencies for the Discord music bot
# Pinned from the working local virtualenv to keep installs reproducible.

# Direct dependencies
discord.py[voice]==2.7.1
lavalink==5.11.0
yt-dlp==2026.3.17

# Transitive runtime dependencies used by discord.py / lavalink / voice support
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
attrs==26.1.0
audioop-lts==0.2.2
cffi==2.0.0
davey==0.1.5
frozenlist==1.8.0
idna==3.11
multidict==6.7.1
propcache==0.4.1
pycparser==3.0
PyNaCl==1.5.0
yarl==1.23.0
